fix builderror poincare

+++ components/ion/include/ion/exam_mode.h
@@ -83,14 +83,14 @@ class Configuration {
                 OMG::BitHelper::numberOfBitsIn<Int>());
 
   bool configurable() const {
-    return OMG::BitHelper::bitAtIndex(m_bits, Bits::Configurable);
+    return OMG::BitHelper::bitAtIndex(m_bits, static_cast<size_t>(Bits::Configurable));
   }
   bool cleared() const {
-    return OMG::BitHelper::bitAtIndex(m_bits, Bits::Cleared);
+    return OMG::BitHelper::bitAtIndex(m_bits, static_cast<size_t>(Bits::Cleared));
   }
   Int data() const {
-    return OMG::BitHelper::bitsBetweenIndexes(m_bits, Bits::DataLast,
-                                              Bits::DataFirst);
+    return OMG::BitHelper::bitsBetweenIndexes(m_bits, static_cast<size_t>(Bits::DataLast),
+                                              static_cast<size_t>(Bits::DataFirst));
   }
 
   Int m_bits;
